proto request type수정

+++ server/src/main/kotlin/com/quid/server/person/gateway/grpc/PersonGrpcController.kt
@@ -16,7 +16,7 @@ class PersonGrpcController(
         request: PersonService.PersonGetRequest,
         responseObserver: StreamObserver<PersonProto>
     ) {
-        personUseCase.getPerson(request.id.toLong())
+        personUseCase.getPerson(request.id)
             .toPersonGrpc()
             .let {
                 responseObserver.onNext(it)
